+++ app/logic/Contacts/EWS/EWSGroup.ts
@@ -1,6 +1,6 @@
+import { ExchangeGroup } from './ExchangeGroup';
 import { Person, ContactEntry } from '../../Abstract/Person';
 import { findPerson } from '../../Abstract/PersonUID';
-import { ExchangeGroup } from './ExchangeGroup';
 import type { EWSAddressbook } from './EWSAddressbook';
 import { EWSCreateItemRequest } from "../../Mail/EWS/Request/EWSCreateItemRequest";
 import { EWSDeleteItemRequest } from "../../Mail/EWS/Request/EWSDeleteItemRequest";
