+++ app/logic/Mail/EWS/EWSAccount.ts
@@ -4,7 +4,7 @@ import { AuthMethod, type Account } from "../../Abstract/Account";
 import type { EMail } from "../EMail";
 import { SpecialFolder, type Folder, type MailShareCombinedPermissions, type MailShareIndividualPermissions } from "../Folder";
 import { EWSFolder, getEWSItem } from "./EWSFolder";
-import { deleteExchangePermissions, setExchangePermissions } from "./ExchangeFolder";
+import { deleteExchangePermissions, setExchangePermissions } from "./ExchangePermission";
 import { EWSCreateItemRequest } from "./Request/EWSCreateItemRequest";
 import type { EWSDeleteItemRequest } from "./Request/EWSDeleteItemRequest";
 import type { EWSUpdateItemRequest } from "./Request/EWSUpdateItemRequest";
